Split Maestro.Contracts
