+++ src/Maestro/Maestro.MergePolicyEvaluation/IMergePolicyInfo.cs
@@ -1,0 +1,13 @@
+// Licensed to the .NET Foundation under one or more agreements.
+// The .NET Foundation licenses this file to you under the MIT license.
+// See the LICENSE file in the project root for more information.
+
+#nullable enable
+namespace Maestro.MergePolicyEvaluation;
+
+public interface IMergePolicyInfo
+{
+    string Name { get; }
+    string DisplayName { get; }
+}
+
