+++ src/Maestro/Maestro.MergePolicyEvaluation/MergePolicyEvaluationResult.cs
@@ -2,28 +2,8 @@
 // The .NET Foundation licenses this file to you under the MIT license.
 // See the LICENSE file in the project root for more information.
 
-using System;
-using System.Collections.Generic;
-using System.Collections.Immutable;
-using System.Linq;
-
-namespace Maestro.Contracts;
-
-public class MergePolicyEvaluationResults
-{
-    public MergePolicyEvaluationResults(IEnumerable<MergePolicyEvaluationResult> results)
-    {
-        Results = results.ToImmutableList();
-    }
-
-    public IImmutableList<MergePolicyEvaluationResult> Results { get; }
-
-    public bool Succeeded => Results.Count > 0 && Results.All(r => r.Status == MergePolicyEvaluationStatus.Success);
-
-    public bool Pending => Results.Any(r => r.Status == MergePolicyEvaluationStatus.Pending);
-
-    public bool Failed => Results.Any(r => r.Status == MergePolicyEvaluationStatus.Failure);
-}
+#nullable enable
+namespace Maestro.MergePolicyEvaluation;
 
 public class MergePolicyEvaluationResult
 {
@@ -50,13 +30,6 @@ public class MergePolicyEvaluationResult
 
     public MergePolicyEvaluationStatus Status { get; }
     public string Title { get; }
-    public string Message { get;}
+    public string Message { get; }
     public IMergePolicyInfo MergePolicyInfo { get; }
-}
-
-public enum MergePolicyEvaluationStatus
-{
-    Pending = 0,
-    Success,
-    Failure,
 }
