+++ src/Maestro/Maestro.MergePolicyEvaluation/MergePolicyEvaluationStatus.cs
@@ -2,12 +2,12 @@
 // The .NET Foundation licenses this file to you under the MIT license.
 // See the LICENSE file in the project root for more information.
 
-using System.Threading.Tasks;
+#nullable enable
+namespace Maestro.MergePolicyEvaluation;
 
-namespace Maestro.Contracts;
-
-public interface IMergePolicyInfo
+public enum MergePolicyEvaluationStatus
 {
-    string Name { get; }
-    string DisplayName { get; }
+    Pending = 0,
+    Success,
+    Failure,
 }
