+++ src/Maestro/SubscriptionActorService/IMergePolicyEvaluator.cs
@@ -6,6 +6,7 @@ using System.Collections.Generic;
 using System.Threading.Tasks;
 using Maestro.Contracts;
 using Maestro.Data.Models;
+using Maestro.MergePolicyEvaluation;
 using Microsoft.DotNet.DarcLib;
 
 namespace SubscriptionActorService;
