+++ src/Maestro/tests/Maestro.ScenarioTests/MaestroScenarioTestBase.cs
@@ -9,6 +9,7 @@ using System.Text.RegularExpressions;
 using System.Threading.Tasks;
 using FluentAssertions;
 using Maestro.Contracts;
+using Maestro.MergePolicyEvaluation;
 using Maestro.ScenarioTests.ObjectHelpers;
 using Microsoft.DotNet.Internal.Testing.Utility;
 using Microsoft.DotNet.Maestro.Client;
