+++ src/Maestro/tests/Maestro.ScenarioTests/ObjectHelpers/SubscriptionBuilder.cs
@@ -3,6 +3,7 @@ using System.Collections.Generic;
 using System.Collections.Immutable;
 using System.Linq;
 using Maestro.Contracts;
+using Maestro.MergePolicyEvaluation;
 using Microsoft.DotNet.Darc;
 using Microsoft.DotNet.Maestro.Client.Models;
 using Newtonsoft.Json.Linq;
