+++ src/Maestro/tests/Maestro.ScenarioTests/ScenarioTests_Subscriptions.cs
@@ -2,6 +2,7 @@ using System;
 using System.Collections.Generic;
 using System.Threading.Tasks;
 using Maestro.Contracts;
+using Maestro.MergePolicyEvaluation;
 using Maestro.ScenarioTests.ObjectHelpers;
 using Microsoft.DotNet.Darc;
 using Microsoft.DotNet.Maestro.Client.Models;
