+++ src/Microsoft.DotNet.Darc/src/DarcLib/IRemoteGitRepo.cs
@@ -3,6 +3,7 @@
 // See the LICENSE file in the project root for more information.
 
 using Maestro.Contracts;
+using Maestro.MergePolicyEvaluation;
 using System.Collections.Generic;
 using System.Threading.Tasks;
