+++ src/Telemetry/TelemetryApplication/ApplicationPackageRoot/ApplicationManifest.xml
@@ -1,4 +1,4 @@
-<?xml version="1.0" encoding="utf-8"?>
+﻿<?xml version="1.0" encoding="utf-8"?>
 <ApplicationManifest xmlns:xsd="http://www.w3.org/2001/XMLSchema" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" ApplicationTypeName="TelemetryApplicationType" ApplicationTypeVersion="1.0.42" xmlns="http://schemas.microsoft.com/2011/01/fabric">
   <Parameters>
     <Parameter Name="AzureDevOpsTimeline_InstanceCount" DefaultValue="1" />
